requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
python-docx==1.2.0
typing_extensions==4.15.0
Werkzeug==3.1.6
